chore(account-lib): refactor addProxy and anonymous txn builders STLX-11678
