+++ modules/account-lib/src/coin/dot/addressInitializationBuilder.ts
@@ -6,17 +6,27 @@ import { TransactionBuilder } from './transactionBuilder';
 import { Transaction } from './transaction';
 import { DecodedSignedTx, DecodedSigningPayload, UnsignedTransaction } from '@substrate/txwrapper-core';
 import { TransactionType } from '../baseCoin';
-import { AddProxyArgs, MethodNames, ProxyType } from './iface';
-import { AddressInitializationSchema } from './txnSchema';
+import { AddAnonymousProxyArgs, AddProxyArgs, MethodNames, ProxyType } from './iface';
+import { AddressInitializationSchema, AnonymousAddressInitializationSchema } from './txnSchema';
 import { BaseAddress } from '../baseCoin/iface';
 
 export class AddressInitializationBuilder extends TransactionBuilder {
   protected _delegate: string;
   protected _proxyType: ProxyType;
   protected _delay: string;
+  protected _index = 0;
 
   constructor(_coinConfig: Readonly<CoinConfig>) {
     super(_coinConfig);
+  }
+
+  /** @inheritDoc */
+  protected buildTransaction(): UnsignedTransaction {
+    if (this._delegate) {
+      return this.buildAddProxyTransaction();
+    } else {
+      return this.buildAnonymousProxyTransaction();
+    }
   }
 
   /**
@@ -26,13 +36,31 @@ export class AddressInitializationBuilder extends TransactionBuilder {
    *
    * @see https://polkadot.js.org/docs/substrate/extrinsics/#proxy
    */
-  protected buildTransaction(): UnsignedTransaction {
+  protected buildAddProxyTransaction(): UnsignedTransaction {
     const baseTxInfo = this.createBaseTxInfo();
     return methods.proxy.addProxy(
       {
         delegate: this._delegate,
         proxyType: this._proxyType,
         delay: this._delay,
+      },
+      baseTxInfo.baseTxInfo,
+      baseTxInfo.options,
+    );
+  }
+
+  /**
+   * Spawn a receive address for the sender
+   *
+   * @return {UnsignedTransaction} an unsigned Dot transaction
+   */
+  protected buildAnonymousProxyTransaction(): UnsignedTransaction {
+    const baseTxInfo = this.createBaseTxInfo();
+    return methods.proxy.anonymous(
+      {
+        proxyType: this._proxyType,
+        index: this._index,
+        delay: parseInt(this._delay, 10),
       },
       baseTxInfo.baseTxInfo,
       baseTxInfo.options,
@@ -54,6 +82,20 @@ export class AddressInitializationBuilder extends TransactionBuilder {
   owner(owner: BaseAddress): this {
     this.validateAddress({ address: owner.address });
     this._delegate = owner.address;
+    return this;
+  }
+
+  /**
+   * Used for disambiguation if multiple calls are made in the same transaction
+   * Use 0 as a default
+   *
+   * @param {number} index
+   *
+   * @returns {AddressInitializationBuilder} This transfer builder.
+   */
+  index(index: number): this {
+    this.validateValue(new BigNumber(index));
+    this._index = index;
     return this;
   }
 
@@ -89,15 +131,20 @@ export class AddressInitializationBuilder extends TransactionBuilder {
 
   /** @inheritdoc */
   validateDecodedTransaction(decodedTxn: DecodedSigningPayload | DecodedSignedTx): void {
+    let validationResult;
     if (decodedTxn.method?.name === MethodNames.AddProxy) {
       const txMethod = decodedTxn.method.args as unknown as AddProxyArgs;
-      const delegate = txMethod.delegate;
-      const proxyType = txMethod.proxyType;
-      const delay = txMethod.delay;
-      const validationResult = AddressInitializationSchema.validate({ delegate, proxyType, delay });
-      if (validationResult.error) {
-        throw new InvalidTransactionError(`Transaction validation failed: ${validationResult.error.message}`);
-      }
+      validationResult = this.validateAddProxyFields(txMethod.delegate, txMethod.proxyType, txMethod.delay);
+    } else if (decodedTxn.method?.name === MethodNames.Anonymous) {
+      const txMethod = decodedTxn.method.args as unknown as AddAnonymousProxyArgs;
+      validationResult = this.validateAnonymousProxyFields(
+        parseInt(txMethod.index, 10),
+        txMethod.proxyType,
+        txMethod.delay,
+      );
+    }
+    if (validationResult.error) {
+      throw new InvalidTransactionError(`Transaction validation failed: ${validationResult.error.message}`);
     }
   }
 
@@ -109,9 +156,14 @@ export class AddressInitializationBuilder extends TransactionBuilder {
       this.owner({ address: txMethod.delegate });
       this.type(txMethod.proxyType);
       this.delay(new BigNumber(txMethod.delay).toString());
+    } else if (this._method?.name === MethodNames.Anonymous) {
+      const txMethod = this._method.args as AddAnonymousProxyArgs;
+      this.index(new BigNumber(txMethod.index).toNumber());
+      this.type(txMethod.proxyType);
+      this.delay(new BigNumber(txMethod.delay).toString());
     } else {
       throw new InvalidTransactionError(
-        `Invalid Transaction Type: ${this._method?.name}. Expected Address initialization`,
+        `Invalid Transaction Type: ${this._method?.name}. Expected ${MethodNames.AddProxy} or ${MethodNames.Anonymous}`,
       );
     }
     return tx;
@@ -120,20 +172,36 @@ export class AddressInitializationBuilder extends TransactionBuilder {
   /** @inheritdoc */
   validateTransaction(_: Transaction): void {
     super.validateTransaction(_);
-    this.validateFields(this._delegate, this._proxyType, this._delay);
+    this.validateFields();
   }
 
-  private validateFields(delegate: string, proxyType: string, delay: string): void {
-    const validationResult = AddressInitializationSchema.validate({
-      delegate,
-      proxyType,
-      delay,
-    });
-
+  private validateFields(): any {
+    let validationResult;
+    if (this._delegate) {
+      validationResult = this.validateAddProxyFields(this._delegate, this._proxyType, this._delay);
+    } else {
+      validationResult = this.validateAnonymousProxyFields(this._index, this._proxyType, this._delay);
+    }
     if (validationResult.error) {
       throw new InvalidTransactionError(
         `AddressInitialization Transaction validation failed: ${validationResult.error.message}`,
       );
     }
+  }
+
+  private validateAddProxyFields(delegate: string, proxyType: string, delay: string): any {
+    return AddressInitializationSchema.validate({
+      delegate,
+      proxyType,
+      delay,
+    });
+  }
+
+  private validateAnonymousProxyFields(index: number, proxyType: string, delay: string): any {
+    return AnonymousAddressInitializationSchema.validate({
+      proxyType,
+      index,
+      delay,
+    });
   }
 }
