+++ modules/account-lib/test/resources/dot/index.ts
@@ -55,6 +55,12 @@ export const rawTx = {
     unsigned:
       '0x9c1d01d472bd6e0f1f92297631938e30edb682208c2cd2698d80cf678c53a69979eb9f0000000000d5012103008c230000070000002b8d4fdbb41f4bc15b8a7ec8ed0687f2a1ae11e0fc2dc6604fa962a9421ae349149799bc9602cb5cf201f3425fb8d253b2d4e61fc119dcab3249f307f594754d',
   },
+  anonymous: {
+    signed:
+      '0xc501840061b18c6dc02ddcabdeac56cb4f21a971cc41cc97640f6f85b073480008c53a0d006f99c126605d8813aeb62ef201a9711dd6edaf941d68dc1e2e84f82354ee39431b0c62d18209b69793066ced404845a12070c56b5ac6cf66c263315c6623da01d5012103001d0400000000000000',
+    unsigned:
+      '0x241d0400000000000000d5012103008c230000070000002b8d4fdbb41f4bc15b8a7ec8ed0687f2a1ae11e0fc2dc6604fa962a9421ae349149799bc9602cb5cf201f3425fb8d253b2d4e61fc119dcab3249f307f594754d',
+  },
   proxy: {
     signed:
       '0xd9028400d472bd6e0f1f92297631938e30edb682208c2cd2698d80cf678c53a69979eb9f004a31bd668659362d682c7d81af0f94455bb1469bab4828b8bfd8b5b439a3fd74bf1240f51aab104e4bfd12cff0eda68df8c01952434a9083f01f250f6e7fc30fd5012103001d0061b18c6dc02ddcabdeac56cb4f21a971cc41cc97640f6f85b073480008c53a0d01000503009f7b0675db59d19b4bd9c8c72eaabba75a9863d02b30115b8b3c3ca5c20f02540bfadb9bbae251',
