+++ modules/account-lib/test/unit/coin/dot/transactionBuilder/addressInitializationBuilder.ts
@@ -5,7 +5,7 @@ import * as DotResources from '../../../../resources/dot';
 import { buildTestConfig } from './base';
 import { ProxyType } from '../../../../../src/coin/dot/iface';
 
-describe('Dot Add Proxy Builder', () => {
+describe('Dot Address Initialization Builder', () => {
   let builder: AddressInitializationBuilder;
 
   const sender = DotResources.accounts.account1;
@@ -15,6 +15,7 @@ describe('Dot Add Proxy Builder', () => {
     const config = buildTestConfig();
     builder = new AddressInitializationBuilder(config);
   });
+
   describe('setter validation', () => {
     it('should validate delay', () => {
       const spy = sinon.spy(builder, 'validateValue');
@@ -33,6 +34,16 @@ describe('Dot Add Proxy Builder', () => {
         (e: Error) => e.message === `The address 'asd' is not a well-formed dot address`,
       );
       should.doesNotThrow(() => builder.owner({ address: sender.address }));
+      assert.calledTwice(spy);
+    });
+
+    it('should validate index', () => {
+      const spy = sinon.spy(builder, 'validateValue');
+      should.throws(
+        () => builder.index(-1),
+        (e: Error) => e.message === 'Value cannot be less than zero',
+      );
+      should.doesNotThrow(() => builder.index(0));
       assert.calledTwice(spy);
     });
   });
@@ -140,6 +151,140 @@ describe('Dot Add Proxy Builder', () => {
       should.deepEqual(txJson.tip, 0);
       should.deepEqual(txJson.transactionVersion, 7);
       should.deepEqual(txJson.chainName, 'Polkadot');
+    });
+  });
+
+  describe('build anonymous proxy creation transaction', () => {
+    it('should build an anonymous proxy transaction', async () => {
+      builder
+        .type(ProxyType.ANY)
+        .delay('0')
+        .index(0)
+        .sender({ address: sender.address })
+        .validity({ firstValid: 3933, maxDuration: 64 })
+        .referenceBlock('0x149799bc9602cb5cf201f3425fb8d253b2d4e61fc119dcab3249f307f594754d')
+        .sequenceId({ name: 'Nonce', keyword: 'nonce', value: 200 })
+        .fee({ amount: 0, type: 'tip' })
+        .version(7);
+      builder.sign({ key: sender.secretKey });
+      const tx = await builder.build();
+      const txJson = tx.toJson();
+      should.deepEqual(txJson.proxyType, ProxyType.ANY);
+      should.deepEqual(txJson.delay, '0');
+      should.deepEqual(txJson.index, '0');
+      should.deepEqual(txJson.sender, sender.address);
+      should.deepEqual(txJson.blockNumber, 3933);
+      should.deepEqual(txJson.referenceBlock, '0x149799bc9602cb5cf201f3425fb8d253b2d4e61fc119dcab3249f307f594754d');
+      should.deepEqual(txJson.genesisHash, '0x2b8d4fdbb41f4bc15b8a7ec8ed0687f2a1ae11e0fc2dc6604fa962a9421ae349');
+      should.deepEqual(txJson.specVersion, 9100);
+      should.deepEqual(txJson.nonce, 200);
+      should.deepEqual(txJson.tip, 0);
+      should.deepEqual(txJson.transactionVersion, 7);
+      should.deepEqual(txJson.chainName, 'Polkadot');
+      should.deepEqual(txJson.eraPeriod, 64);
+    });
+
+    it('should build an unsigned anonymous proxy transaction', async () => {
+      builder
+        .type(ProxyType.ANY)
+        .delay('0')
+        .index(0)
+        .sender({ address: sender.address })
+        .validity({ firstValid: 3933, maxDuration: 64 })
+        .referenceBlock('0x149799bc9602cb5cf201f3425fb8d253b2d4e61fc119dcab3249f307f594754d')
+        .sequenceId({ name: 'Nonce', keyword: 'nonce', value: 200 })
+        .fee({ amount: 0, type: 'tip' })
+        .version(7);
+      const tx = await builder.build();
+      const txJson = tx.toJson();
+      should.deepEqual(txJson.proxyType, ProxyType.ANY);
+      should.deepEqual(txJson.delay, '0');
+      should.deepEqual(txJson.index, '0');
+      should.deepEqual(txJson.sender, sender.address);
+      should.deepEqual(txJson.blockNumber, 3933);
+      should.deepEqual(txJson.referenceBlock, '0x149799bc9602cb5cf201f3425fb8d253b2d4e61fc119dcab3249f307f594754d');
+      should.deepEqual(txJson.genesisHash, '0x2b8d4fdbb41f4bc15b8a7ec8ed0687f2a1ae11e0fc2dc6604fa962a9421ae349');
+      should.deepEqual(txJson.specVersion, 9100);
+      should.deepEqual(txJson.nonce, 200);
+      should.deepEqual(txJson.tip, 0);
+      should.deepEqual(txJson.transactionVersion, 7);
+      should.deepEqual(txJson.chainName, 'Polkadot');
+      should.deepEqual(txJson.eraPeriod, 64);
+    });
+
+    it('should default to building anonymous proxy if owner is not called', async () => {
+      builder
+        .type(ProxyType.ANY)
+        .delay('0')
+        .sender({ address: sender.address })
+        .validity({ firstValid: 3933, maxDuration: 64 })
+        .referenceBlock('0x149799bc9602cb5cf201f3425fb8d253b2d4e61fc119dcab3249f307f594754d')
+        .sequenceId({ name: 'Nonce', keyword: 'nonce', value: 200 })
+        .fee({ amount: 0, type: 'tip' })
+        .version(7);
+      builder.sign({ key: sender.secretKey });
+      const tx = await builder.build();
+      const txJson = tx.toJson();
+      should.deepEqual(txJson.proxyType, ProxyType.ANY);
+      should.deepEqual(txJson.delay, '0');
+      should.deepEqual(txJson.index, '0');
+      should.deepEqual(txJson.sender, sender.address);
+      should.deepEqual(txJson.blockNumber, 3933);
+      should.deepEqual(txJson.referenceBlock, '0x149799bc9602cb5cf201f3425fb8d253b2d4e61fc119dcab3249f307f594754d');
+      should.deepEqual(txJson.genesisHash, '0x2b8d4fdbb41f4bc15b8a7ec8ed0687f2a1ae11e0fc2dc6604fa962a9421ae349');
+      should.deepEqual(txJson.specVersion, 9100);
+      should.deepEqual(txJson.nonce, 200);
+      should.deepEqual(txJson.tip, 0);
+      should.deepEqual(txJson.transactionVersion, 7);
+      should.deepEqual(txJson.chainName, 'Polkadot');
+      should.deepEqual(txJson.eraPeriod, 64);
+    });
+
+    it('should build from raw signed tx', async () => {
+      builder.from(DotResources.rawTx.anonymous.signed);
+      builder
+        .validity({ firstValid: 8975007, maxDuration: 64 })
+        .referenceBlock('0x9ed0c8ee5fdc375ee57f79591d7d0db4d7cd2aa0e5403a2ed84edf0f859e3f05')
+        .version(7);
+      const tx = await builder.build();
+      const txJson = tx.toJson();
+      should.deepEqual(txJson.proxyType, ProxyType.ANY);
+      should.deepEqual(txJson.index, '0');
+      should.deepEqual(txJson.delay, '0');
+      should.deepEqual(txJson.sender, sender.address);
+      should.deepEqual(txJson.blockNumber, 8975007);
+      should.deepEqual(txJson.referenceBlock, '0x9ed0c8ee5fdc375ee57f79591d7d0db4d7cd2aa0e5403a2ed84edf0f859e3f05');
+      should.deepEqual(txJson.genesisHash, '0x2b8d4fdbb41f4bc15b8a7ec8ed0687f2a1ae11e0fc2dc6604fa962a9421ae349');
+      should.deepEqual(txJson.specVersion, 9100);
+      should.deepEqual(txJson.nonce, 200);
+      should.deepEqual(txJson.tip, 0);
+      should.deepEqual(txJson.transactionVersion, 7);
+      should.deepEqual(txJson.chainName, 'Polkadot');
+      should.deepEqual(txJson.eraPeriod, 64);
+    });
+
+    it('should build from raw unsigned tx', async () => {
+      builder.from(DotResources.rawTx.anonymous.unsigned);
+      builder
+        .validity({ firstValid: 8975007, maxDuration: 64 })
+        .referenceBlock('0x9ed0c8ee5fdc375ee57f79591d7d0db4d7cd2aa0e5403a2ed84edf0f859e3f05')
+        .sender({ address: sender.address })
+        .sign({ key: sender.secretKey });
+      const tx = await builder.build();
+      const txJson = tx.toJson();
+      should.deepEqual(txJson.proxyType, ProxyType.ANY);
+      should.deepEqual(txJson.index, '0');
+      should.deepEqual(txJson.delay, '0');
+      should.deepEqual(txJson.sender, sender.address);
+      should.deepEqual(txJson.blockNumber, 8975007);
+      should.deepEqual(txJson.referenceBlock, '0x9ed0c8ee5fdc375ee57f79591d7d0db4d7cd2aa0e5403a2ed84edf0f859e3f05');
+      should.deepEqual(txJson.genesisHash, '0x2b8d4fdbb41f4bc15b8a7ec8ed0687f2a1ae11e0fc2dc6604fa962a9421ae349');
+      should.deepEqual(txJson.specVersion, 9100);
+      should.deepEqual(txJson.nonce, 200);
+      should.deepEqual(txJson.tip, 0);
+      should.deepEqual(txJson.transactionVersion, 7);
+      should.deepEqual(txJson.chainName, 'Polkadot');
+      should.deepEqual(txJson.eraPeriod, 64);
     });
   });
 });
